+++ lib/easypost/services/billing.rb
@@ -62,15 +62,15 @@ class EasyPost::Services::Billing < EasyPost::Services::Service
     end
 
     payment_method_id = payment_methods[payment_method_to_use]['id']
+    payment_method_object_type = payment_methods[payment_method_to_use]['object']
 
-    unless payment_method_id.nil?
-      if payment_method_id.start_with?('card_')
-        endpoint = '/credit_cards'
-      elsif payment_method_id.start_with?('bank_')
-        endpoint = '/bank_accounts'
-      else
-        raise EasyPost::Errors::InvalidObjectError.new(error_string)
-      end
+    if payment_method_object_type == 'CreditCard'
+
+      endpoint = '/credit_cards'
+    elsif payment_method_object_type == 'BankAccount'
+      endpoint = '/bank_accounts'
+    else
+      raise EasyPost::Errors::InvalidObjectError.new(error_string)
     end
 
     [endpoint, payment_method_id]