+++ spec/billing_spec.rb
@@ -11,10 +11,10 @@ describe EasyPost::Services::Billing do
                                              .and_return({
                                                            'id' => 'cust_thisisdummydata',
                                                            'object' => 'PaymentMethods', 'primary_payment_method' =>
-                                                             { 'id' => 'card_123', 'object' => 'CreditCard' },
+                                                             { 'id' => 'pm_123', 'object' => 'CreditCard' },
                                                          },
                                              )
-      allow(client).to receive(:make_request).with(:post, '/credit_cards/card_123/charges', { amount: '2000' })
+      allow(client).to receive(:make_request).with(:post, '/credit_cards/pm_123/charges', { amount: '2000' })
       credit_card = client.billing.fund_wallet('2000', 'primary')
 
       expect(credit_card).to eq(true)
@@ -27,10 +27,33 @@ describe EasyPost::Services::Billing do
                                              .and_return({
                                                            'id' => 'cust_thisisdummydata',
                                                            'object' => 'PaymentMethods', 'primary_payment_method' =>
-                                                             { 'id' => 'card_123', 'object' => 'CreditCard' },
+                                                             { 'id' => 'pm_123', 'object' => 'CreditCard' },
                                                          },
                                              )
-      allow(client).to receive(:make_request).with(:delete, '/credit_cards/card_123')
+      allow(client).to receive(:make_request).with(:delete, '/credit_cards/pm_123')
+
+      deleted_credit_card = client.billing.delete_payment_method('primary')
+
+      expect(deleted_credit_card).to eq(true)
+    end
+  end
+
+  describe '.get_payment_method_info' do
+    it 'get payment method type by object type' do
+      allow(client).to receive(:make_request).with(:get, '/payment_methods')
+                                             .and_return({
+                                                           'id' => 'cust_thisisdummydata',
+                                                           'object' => 'PaymentMethods',
+                                                           'primary_payment_method' =>
+                                                             { 'id' => 'pm_123', 'object' => 'CreditCard' },
+                                                           'secondary_payment_method' =>
+                                                             { 'id' => 'pm_456', 'object' => 'BankAccount' },
+                                                         },
+                                                        )
+
+      # get_payment_method_info is private, can test it via delete
+      # will pass if get_payment_method_info returns ['credit_cards', 'pm_123'], fail otherwise
+      allow(client).to receive(:make_request).with(:delete, '/credit_cards/pm_123')
 
       deleted_credit_card = client.billing.delete_payment_method('primary')